+++ include/logrin/Logger.h
@@ -59,10 +59,10 @@ struct VIOLET_API LogEntry final {
 private:
     friend struct Logger;
 
-    VIOLET_EXPLICIT LogEntry(Logger* logger, LogRecord record, bool emit = true) noexcept;
+    VIOLET_EXPLICIT LogEntry(const Logger* logger, LogRecord record, bool emit = true) noexcept;
 
+    const Logger* n_parent; ///< pointer to the parent logger
     LogRecord n_record; ///< the record that this entry is used for
-    Logger* n_parent; ///< pointer to the parent logger
     bool n_emit = false;
 };
 
@@ -142,7 +142,7 @@ struct VIOLET_API Logger final {
     [[nodiscard]] auto Name() const noexcept -> violet::Str;
 
     /// Logs a new record and allows manipulation of the entry itself.
-    auto Log(LogRecord record) noexcept -> LogEntry;
+    [[maybe_unused]] auto Log(LogRecord record) const noexcept -> LogEntry; // NOLINT(modernize-use-nodiscard)
 
     /// Logs a message at a specified log level.
     ///
@@ -152,8 +152,8 @@ struct VIOLET_API Logger final {
     /// @param loc source location, defaults to the caller.
     /// @returns a `LogEntry` for further attribute attachment.
     template<std::convertible_to<violet::Str> Msg>
-    auto Log(LogLevel level, Msg&& message, const std::source_location& loc = std::source_location::current()) noexcept
-        -> LogEntry
+    auto Log(LogLevel level, Msg&& message,
+        const std::source_location& loc = std::source_location::current()) const noexcept -> LogEntry
     {
         if (!this->Enabled(level)) {
             return LogEntry(nullptr, LogRecord{}, false);
@@ -164,42 +164,48 @@ struct VIOLET_API Logger final {
 
     /// Convenience method: logs a Trace-level message.
     template<std::convertible_to<violet::Str> Msg>
-    auto Trace(Msg&& message, const std::source_location& loc = std::source_location::current()) noexcept -> LogEntry
+    auto Trace(Msg&& message, const std::source_location& loc = std::source_location::current()) const noexcept
+        -> LogEntry
     {
         return this->Log(LogRecord::Now(LogLevel::Trace, VIOLET_FWD(Msg, message), loc));
     }
 
     /// Convenience method: logs a Debug-level message.
     template<std::convertible_to<violet::Str> Msg>
-    auto Debug(Msg&& message, const std::source_location& loc = std::source_location::current()) noexcept -> LogEntry
+    auto Debug(Msg&& message, const std::source_location& loc = std::source_location::current()) const noexcept
+        -> LogEntry
     {
         return this->Log(LogRecord::Now(LogLevel::Debug, VIOLET_FWD(Msg, message), loc));
     }
 
     /// Convenience method: logs a Info-level message.
     template<std::convertible_to<violet::Str> Msg>
-    auto Info(Msg&& message, const std::source_location& loc = std::source_location::current()) noexcept -> LogEntry
+    auto Info(Msg&& message, const std::source_location& loc = std::source_location::current()) const noexcept
+        -> LogEntry
     {
         return this->Log(LogRecord::Now(LogLevel::Info, VIOLET_FWD(Msg, message), loc));
     }
 
     /// Convenience method: logs a Warning-level message.
     template<std::convertible_to<violet::Str> Msg>
-    auto Warn(Msg&& message, const std::source_location& loc = std::source_location::current()) noexcept -> LogEntry
+    auto Warn(Msg&& message, const std::source_location& loc = std::source_location::current()) const noexcept
+        -> LogEntry
     {
         return this->Log(LogRecord::Now(LogLevel::Warning, VIOLET_FWD(Msg, message), loc));
     }
 
     /// Convenience method: logs a Error-level message.
     template<std::convertible_to<violet::Str> Msg>
-    auto Error(Msg&& message, const std::source_location& loc = std::source_location::current()) noexcept -> LogEntry
+    auto Error(Msg&& message, const std::source_location& loc = std::source_location::current()) const noexcept
+        -> LogEntry
     {
         return this->Log(LogRecord::Now(LogLevel::Error, VIOLET_FWD(Msg, message), loc));
     }
 
     /// Convenience method: logs a Fatal-level message.
     template<std::convertible_to<violet::Str> Msg>
-    auto Fatal(Msg&& message, const std::source_location& loc = std::source_location::current()) noexcept -> LogEntry
+    auto Fatal(Msg&& message, const std::source_location& loc = std::source_location::current()) const noexcept
+        -> LogEntry
     {
         return this->Log(LogRecord::Now(LogLevel::Fatal, VIOLET_FWD(Msg, message), loc));
     }